Introduce tag transformer
Migrate tag-specific logic from array to new transformer.
Migrate related tests.
