+++ src/functions/transformers/index.js
@@ -1,0 +1,9 @@
+import {
+  determineExpertiseFromTags,
+  stripExpertiseFromTags
+} from './transformTags';
+
+export {
+  determineExpertiseFromTags,
+  stripExpertiseFromTags
+};
