+++ src/functions/utils/array.js
@@ -1,20 +1,3 @@
-import { EXPERTISE_LEVELS } from 'shared';
-
-/**
- * Given an array of tags, determine the expertise level.
- */
-export const determineExpertiseFromTags = tags =>
-  tags.reduce((expertise, tag) =>
-    EXPERTISE_LEVELS.includes(tag) ? tag : expertise,
-  EXPERTISE_LEVELS[1]
-  );
-
-/**
- * Given an array of tags, strip the expertise level.
- */
-export const stripExpertiseFromTags = tags =>
-  tags.filter(tag => !EXPERTISE_LEVELS.includes(tag));
-
 /** Chunks an array into smaller arrays of a specified size. */
 export const chunk = (arr, size) =>
   Array.from({ length: Math.ceil(arr.length / size) }, (v, i) =>
@@ -27,8 +10,8 @@ export const transformSnippetIndex = edges =>
     .map(edge => edge.node)
     .map(node => ({
       title: node.title,
-      expertise: node.expertise,
-      primaryTag: node.tags.primary,
+      expertise: node.expertise)
+      primaryTag: node.tags.primary)
       language: node.language.long,
       description: node.html.description.trim(),
       url: node.slug,
